+++ rv/modules/distortion.py
@@ -14,7 +14,12 @@ class Distortion(Module):
 
     class Type(Enum):
         lim = 0
+        clipping = 0
         sat = 1
+        foldback = 1
+        foldback2 = 2
+        foldback3 = 3
+        overflow = 4
 
     volume = Controller((0, 256), 128)
     type = Controller(Type, Type.lim)